be/fix: 리뷰, 기록 난이도 타입 수정

+++ backend/src/main/java/com/moreroom/domain/review/entity/Review.java
@@ -30,9 +30,8 @@ public class Review extends BaseEntity {
     @Column(nullable = false)
     private Integer themeId;
     private String content;
-    // 별점 Integer 맞나요? float 아니고?
     @Column(nullable = false)
-    private Integer score;
+    private float score;
     @LastModifiedDate
     private LocalDateTime updatedAt;
     @Builder.Default
